chore: move core test files to src/server/tests/
