+++ src/server/tests/base-path-logout-integration.test.ts
@@ -1,10 +1,10 @@
 import { RequestCookies, ResponseCookies } from "@edge-runtime/cookies";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { Auth0Client } from "./client.js";
-import { deleteChunkedCookie, deleteCookie } from "./cookies.js";
-import { StatelessSessionStore } from "./session/stateless-session-store.js";
-import { TransactionStore } from "./transaction-store.js";
+import { Auth0Client } from "../client.js";
+import { deleteChunkedCookie, deleteCookie } from "../cookies.js";
+import { StatelessSessionStore } from "../session/stateless-session-store.js";
+import { TransactionStore } from "../transaction-store.js";
 
 describe("Base path and cookie configuration tests", () => {
   let originalEnv: NodeJS.ProcessEnv;
