+++ src/server/tests/get-access-token-mrrt.test.ts
@@ -13,10 +13,10 @@ import {
   vi
 } from "vitest";
 
-import { createNextHeadersMock } from "../test-fixtures/mocks.js";
-import { SessionData } from "../types/index.js";
-import { DEFAULT_SCOPES } from "../utils/constants.js";
-import { Auth0Client } from "./client.js";
+import { createNextHeadersMock } from "../../test-fixtures/mocks.js";
+import { SessionData } from "../../types/index.js";
+import { DEFAULT_SCOPES } from "../../utils/constants.js";
+import { Auth0Client } from "../client.js";
 
 vi.mock("next/headers.js", () => createNextHeadersMock());
