+++ src/server/tests/logout-cookie-domain.flow.test.ts
@@ -1,7 +1,7 @@
 import { ResponseCookies } from "@edge-runtime/cookies";
 import { describe, expect, it } from "vitest";
 
-import { StatelessSessionStore } from "./session/stateless-session-store.js";
+import { StatelessSessionStore } from "../session/stateless-session-store.js";
 
 describe("Session cookie domain deletion bug", () => {
   it("should delete session cookies with domain when AUTH0_COOKIE_DOMAIN is set", () => {
