+++ src/server/tests/organizations.test.ts
@@ -1,10 +1,10 @@
 import { NextRequest } from "next/server.js";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { getDefaultRoutes } from "../test-fixtures/defaults.js";
-import { AuthClient } from "./auth-client.js";
-import { StatelessSessionStore } from "./session/stateless-session-store.js";
-import { TransactionStore } from "./transaction-store.js";
+import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
+import { AuthClient } from "../auth-client.js";
+import { StatelessSessionStore } from "../session/stateless-session-store.js";
+import { TransactionStore } from "../transaction-store.js";
 
 const DEFAULT = {
   domain: "test.auth0.com",
